[AIBRAIN-18] 问题： 1. 支持新的一维卡尔曼滤波算法

+++ Project/ds_twr_init/kalman/kalman.h
@@ -23,15 +23,15 @@ extern float KalMan_Update(double *Z);
 // Aibrain code
 // 1. 结构体类型定义
 typedef struct {
-    float LastP;    // 上次估算协方差 初始化值为0.02
-    float Now_P;    // 当前估算协方差 初始化值为0
-    float out;  // 卡尔曼滤波器输出 初始化值为0
-    float Kg;   // 卡尔曼增益 初始化值为0
-    float Q;    // 过程噪声协方差 初始化值为0.001
-    float R;    // 观测噪声协方差 初始化值为0.543
+    double LastP;    // 上次估算协方差 初始化值为0.02
+    double Now_P;    // 当前估算协方差 初始化值为0
+    double out;  // 卡尔曼滤波器输出 初始化值为0
+    double Kg;   // 卡尔曼增益 初始化值为0
+    double Q;    // 过程噪声协方差 初始化值为0.001
+    double R;    // 观测噪声协方差 初始化值为0.543
 } KFP; // Kalman Filter parameter
 
-float kalmanFilter(KFP *kfp, float input);
+double kalmanFilter(KFP *kfp, double input);
 
 #endif
